ETC/BE/requirements.txt
Django==3.1
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kerascv==0.0.40
librosa==0.8.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
opencv-python==4.4.0.40
pandas==1.0.5
Pillow==7.2.0
scikit-learn @ file:///C:/ci/scikit-learn_1592847564598/work
scikit-surprise==1.1.1
scipy @ file:///C:/ci/scipy_1592916958183/work
seaborn==0.10.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.8.0
tqdm==4.47.0